Proposition of fix for bug: #147
As far as I understand, Hector would like a warning if an observable hit is tracked and no compression mode is activated.

I propose the following solution:
When I encounter the first observable, I check whether there is at least one compression mode activated, and issue a warning if this is not the case.

Issues:
 -> It breaks the progress bar.
 -> I have no formatter in rule_interpretter update_edges so I have used std_formatter as a temporary solution.

Could you check whether this is OK, and propose alternative solution if this is not the case ?