Fix #644: Use `Buffer.t` instead of `Bi_outbuf.t`
This should make Kappa compatible with the latest version of yojson.

Since https://github.com/ocaml-community/yojson/pull/74, yojson does
not rely on biniou buffers.

The main drawback seems to be that `JsonUtil.write_to_channel` now
writes everything to the memory (in a `Buffer.t`) before dumping the
buffer to the channel.